requirements.txt
# Orchestration & ETL
apache-airflow
minio
pandas
numpy
requests
beautifulsoup4
html5lib
lxml

# Database
sqlalchemy
psycopg2-binary
boto3

# Dev & Utils
python-dotenv

# Jupyter / Dev environment
jupyter_core
jupyter_client
ipykernel
ipython

# Logging & CLI
colorlog
rich
kaggle
pyarrow

# MCP Server
fastmcp

# Testing & Quality
pytest
pytest-spark

# Big Data & Storage
pyspark>=3.3.0
delta-spark